// src/models/response/OpportunityListitemAux.tsx
// Generated by https://quicktype.io

export interface OpportunityListitemAux {
        id:number,
        lote:string,
        item:string,
        valorFechado:string | null,
        valorTotal:string | null,
        ganador:string| null,
        producto:string | null,
        participacion:string | null,
        posicion:string| null,
        collapsed:boolean
}